requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0
scikit-learn>=0.24.0
tensorflow>=2.5.0
keras>=2.5.0
